experiments/python-plugin/requirements.txt
grpcio==1.66.0
grpcio-health-checking==1.59.0
grpcio-tools==1.66.0
protobuf==5.27.3
setuptools==73.0.1
